backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.6
supabase==2.7.3
httpx==0.27.2
PyJWT==2.9.0

